feat: Complete HR and Lecturer features with decimal hourly rate support
BREAKING CHANGE: HourlyRate changed from int to decimal across the system

Features Added:
- HR can create lecturer accounts with CreateLecturer action
- HR can bulk import lecturers via CSV (creates AppUser accounts)
- Lecturers have hourly rate auto-filled from their profile (read-only)
- Auto-calculation of claim amounts based on hours × hourly rate
- Monthly hour limit validation (180 hours per month)
- HomeController now uses logged-in user instead of first lecturer

Technical Changes:
- Changed Claim.HourlyRate from int to decimal for precision
- Changed ClaimSubmissionViewModel.HourlyRate to decimal
- Updated ValidHourlyRateAttribute to handle decimal types
- Added decimal casting to ~50+ calculations across controllers and views
- Injected UserManager into ClaimsController and HomeController
- Fixed TotalAmount calculation in ClaimSubmissionViewModel

Controllers Updated:
- HRController: CreateLecturer, ImportLecturersCSV
- ClaimsController: Pre-fill hourly rate, monthly validation
- HomeController: Use logged-in user
- CoordinatorController, ManagerController, AdminController, LecturerController

Views Updated:
- Made hourly rate field readonly in Claims/Create
- Fixed decimal calculations in 40+ view files
- Added CreateLecturer view for HR
- Updated ManageLecturers with 'Add Lecturer' button

+++ ST10440112-PROG6212-CCMS/Controllers/Admin/ManagerController.cs
@@ -31,7 +31,7 @@ namespace ST10440112_PROG6212_CCMS.Controllers.Admin
                 var rejectedClaims = await _context.Claims.CountAsync(c => c.ClaimStatus == "Rejected");
                 var totalAmount = await _context.Claims
                     .Where(c => c.ClaimStatus == "Approved")
-                    .SumAsync(c => c.TotalHours * c.HourlyRate);
+                    .SumAsync(c => (decimal)c.TotalHours * c.HourlyRate);
 
                 ViewBag.TotalClaims = totalClaims;
                 ViewBag.PendingClaims = pendingClaims;
